pyproject.toml
[project]
name = "smart-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocache[redis]>=0.12.3",
    "beautifulsoup4>=4.14.2",
    "ddgs>=9.6.1",
    "fastapi>=0.120.0",
    "httpx>=0.28.1",
    "instructor>=1.11.3",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-huggingface>=0.1.0",
    "langchain-qdrant>=1.1.0",
    "langchain-tavily>=0.2.12",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.1",
    "langmem>=0.0.30",
    "markdownify>=1.2.0",
    "omegaconf>=2.3.0",
    "psycopg-binary>=3.2.13",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pypdf>=6.1.3",
    "qdrant-client>=1.15.1",
    "rank-bm25>=0.2.2",
    "rich>=14.2.0",
    "sentence-transformers>=5.1.2",
    "slowapi>=0.1.9",
    "streamlit>=1.41.1",
    "together>=1.5.29",
    "tokenizers>=0.22.1",
    "uvicorn>=0.38.0",
    "uvloop>=0.22.1",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = [
    "grandalf>=0.8",
    "jupyter>=1.1.1",
    "nb-black-formatter>=1.0.1",
    "pandas>=2.3.3",
    "polars>=1.34.0",
    "prek>=0.2.11",
    "ruff>=0.14.1",
    "tabulate>=0.9.0",
    "ty>=0.0.1a26",
]
